# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Currículo Fácil",
    "description": "Uma ferramenta desenvolvida para gerar currículos de impacto ou simples de forma simples e rápida",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "gilamar@kemiocode.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
...
  "paths": {
...
    "/networks/{networkId}": {
      "get": {
        "tags": [
          "Networks"
        ],
        "summary": "Retorna um network especifico pelo seu Id, desde que seja de propriedade do usuário",
        "description": "Rota privada que só pode ser acessado pelo mesmo id de quem estiver logado",
        "operationId": "network",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "description": "Id do network",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um objeto network",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Network"
                }
              }
            }
          },
          "401": {
            "description": "Usuário logado diferente do que se busca"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "put": {
        "tags": [
          "Networks"
        ],
        "summary": "Atualiza um network especifico pelo seu Id, desde que seja de propriedade do usuário",
        "description": "Rota privada só com acesso logado",
        "operationId": "editNetwork",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "description": "Id do network",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Os campos created e updated são preenchidos automanticamente",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetworkEditFormData"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/NetworkEditFormData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Objeto network",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Network"
                }
              }
            }
          },
          "204": {
            "description": "Usuário envio algum campo do formulário em branco"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "Networks"
        ],
        "summary": "Deleta o um network especifico pelo seu Id, desde que seja de propriedade do usuário",
        "description": "Rota privada só com acesso logado",
        "operationId": "deleteNetwork",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "description": "Id do network",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "network excluido com sucesso"
          },
          "401": {
            "description": "Usuário logado diferente do que se busca o Id"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/languages/types": {
      "get": {
        "tags": [
          "Languages"
        ],
        "summary": "Retorna todos os tipos de languages",
        "description": "Rota privada listar todos os languages types do sistema",
        "operationId": "languagesTypes",
        "responses": {
          "200": {
            "description": "Array de objetos language type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageType"
                }
              }
            }
          },
          "401": {
            "description": "Usuário não logado no sistema"
          }
        }
      }
    },
...
    "/languages/{languageLevelId}": {
      "get": {
        "tags": [
          "Languages"
        ],
        "summary": "Retorna um language especifico pelo seu Id, desde que seja de propriedade do usuário",
        "description": "Rota privada que só pode ser acessado pelo mesmo id de quem estiver logado",
        "operationId": "language",
        "parameters": [
          {
            "name": "languageLevelId",
            "in": "path",
            "description": "Id do level language",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um objeto language",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          },
          "401": {
            "description": "Usuário logado diferente do que se busca"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "put": {
        "tags": [
          "Languages"
        ],
        "summary": "Atualiza um language especifico pelo seu Id, desde que seja de propriedade do usuário",
        "description": "Rota privada só com acesso logado",
        "operationId": "editLanguage",
        "parameters": [
          {
            "name": "languageLevelId",
            "in": "path",
            "description": "Id do language",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Os campos created e updated são preenchidos automanticamente",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguageEditFormData"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LanguageEditFormData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Objeto language",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          },
          "204": {
            "description": "Usuário envio algum campo do formulário em branco"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "Languages"
        ],
        "summary": "Deleta o um language especifico pelo seu Id, desde que seja de propriedade do usuário",
        "description": "Rota privada só com acesso logado",
        "operationId": "deleteLanguage",
        "parameters": [
          {
            "name": "languageLevelId",
            "in": "path",
            "description": "Id do language",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "language excluido com sucesso"
          },
          "401": {
            "description": "Usuário logado diferente do que se busca o Id"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Network": {
        "type": "object",
        "required": [
          "id",
          "network",
          "networkTypeId",
          "userId",
          "created",
          "updated"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "network": {
            "type": "string"
          },
          "networkTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NetworkEditFormData": {
        "type": "object",
        "required": [
          "link",
          "networkTypeId"
        ],
        "properties": {
          "link": {
            "type": "string"
          },
          "networkTypeId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Language": {
        "type": "object",
        "required": [
          "id",
          "network",
          "networkTypeId",
          "userId",
          "created",
          "updated"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "network": {
            "type": "string"
          },
          "networkTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LanguageEditFormData": {
        "type": "object",
        "required": [
          "stars",
          "languageId"
        ],
        "properties": {
          "stars": {
            "type": "integer",
            "format": "int64"
          },
          "languageId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "LanguageType": {
        "type": "object",
        "required": [
          "id",
          "network",
          "created",
          "updated"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}